requirements.txt
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
schedule==1.2.1
pyarrow==14.0.2
openpyxl==3.1.2
scikit-learn==1.3.2
xgboost==2.0.3
matplotlib==3.8.2
seaborn==0.13.0
jupyter==1.0.0
tqdm==4.66.1
loguru==0.7.2
